# platform/api-contracts/inference-gateway.openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunBudget": {
        "additionalProperties": false,
        "properties": {
          "cost_limit_usd": {
            "default": 5.0,
            "exclusiveMinimum": 0.0,
            "maximum": 1000000.0,
            "title": "Cost Limit Usd",
            "type": "number"
          },
          "token_limit": {
            "default": 10000,
            "exclusiveMinimum": 0.0,
            "maximum": 1000000000.0,
            "title": "Token Limit",
            "type": "integer"
          },
          "workflow": {
            "default": "",
            "maxLength": 128,
            "title": "Workflow",
            "type": "string"
          }
        },
        "title": "RunBudget",
        "type": "object"
      },
...
      "ToolCall": {
        "additionalProperties": false,
        "properties": {
          "arguments": {
            "additionalProperties": true,
            "title": "Arguments",
            "type": "object"
          }
        },
        "required": [
          "arguments"
        ],
        "title": "ToolCall",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "description": "API key header accepted by gateway middleware when API key authentication is enabled.",
        "in": "header",
        "name": "X-API-Key",
        "type": "apiKey"
      },
      "BearerAuth": {
        "description": "Bearer token accepted by gateway middleware when API key authentication is enabled.",
        "scheme": "bearer",
        "type": "http"
      },
      "CsrfToken": {
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      },
      "GitHubSignature": {
        "description": "GitHub HMAC-SHA256 of the raw body using the team's secret. Also send X-GitHub-Delivery. The signed body fingerprint protects against replay.",
        "in": "header",
        "name": "X-Hub-Signature-256",
        "type": "apiKey"
      },
      "SessionCookieAuth": {
        "description": "Opaque console session; non-GET requests also require the double-submit CSRF token.",
        "in": "cookie",
        "name": "aw_session",
        "type": "apiKey"
      },
      "WebhookSignature": {
        "description": "sha256=<HMAC-SHA256(team secret, timestamp.delivery.path.raw_body)>; also send X-AW-Timestamp (Unix seconds, within 5 minutes) and X-AW-Delivery (unique ID).",
        "in": "header",
        "name": "X-AW-Signature",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "OpenAI-compatible private inference gateway with sandbox traceability, admission controls, budget enforcement, redacted audit events, and runtime routing to Ollama or vLLM.",
    "title": "AgentWorkflows Inference Gateway",
    "version": "0.5.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/tools": {
      "get": {
        "operationId": "tools_v1_tools_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Tools V1 Tools Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          },
          {
            "SessionCookieAuth": []
          }
        ],
        "summary": "List this team's approved workflow tools",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/tools/{tool}/call": {
      "post": {
        "operationId": "call_tool_v1_tools__tool__call_post",
        "parameters": [
          {
            "in": "path",
            "name": "tool",
            "required": true,
            "schema": {
              "title": "Tool",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ToolCall"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Call Tool V1 Tools  Tool  Call Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          },
          {
            "CsrfToken": [],
            "SessionCookieAuth": []
          }
        ],
        "summary": "Execute an approved tool through gateway governance",
        "tags": [
          "workflows"
        ]
      }
    },
...
    "/v1/workflow-policies": {
      "get": {
        "operationId": "workflow_policies_v1_workflow_policies_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Workflow Policies V1 Workflow Policies Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          },
          {
            "SessionCookieAuth": []
          }
        ],
        "summary": "Discover this team's workflow policies",
        "tags": [
          "workflows"
        ]
      }
    },
...
    "/v1/workflow-runs/{run_id}": {
      "get": {
        "operationId": "run_usage_v1_workflow_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Run Usage V1 Workflow Runs  Run Id  Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          },
          {
            "SessionCookieAuth": []
          }
        ],
        "summary": "Inspect this team's run budget",
        "tags": [
          "workflows"
        ]
      },
      "put": {
        "operationId": "initialize_run_v1_workflow_runs__run_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunBudget"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Initialize Run V1 Workflow Runs  Run Id  Put",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ApiKeyAuth": []
          },
          {
            "CsrfToken": [],
            "SessionCookieAuth": []
          }
        ],
        "summary": "Initialize an immutable per-run budget",
        "tags": [
          "workflows"
        ]
      }
    },
...
  },
...
}